+++ code-samples/additional-info/rest-api/13/index.js
@@ -1,6 +1,5 @@
 const { Client, Intents, MessageEmbed } = require('discord.js');
 const fetch = require('node-fetch');
-const querystring = require('querystring');
 
 const client = new Client({ intents: [Intents.FLAGS.GUILDS] });
 
@@ -13,20 +12,22 @@ client.once('ready', () => {
 client.on('interactionCreate', async interaction => {
 	if (!interaction.isCommand()) return;
 
-	const { commandName: command } = interaction;
+	const { commandName } = interaction;
 
-	if (command === 'cat') {
+	if (commandName === 'cat') {
+		await interaction.deferReply();
 		const { file } = await fetch('https://aws.random.cat/meow').then(response => response.json());
 
-		interaction.reply({ files: [file] });
-	} else if (command === 'urban') {
+		interaction.editReply({ files: [file] });
+	} else if (commandName === 'urban') {
+		await interaction.deferReply();
 		const term = interaction.options.getString('term');
-		const query = querystring.stringify({ term });
+		const query = new URLSearchParams({ term });
 
 		const { list } = await fetch(`https://api.urbandictionary.com/v0/define?${query}`).then(response => response.json());
 
 		if (!list.length) {
-			return interaction.reply(`No results found for **${term}**.`);
+			return interaction.editReply(`No results found for **${term}**.`);
 		}
 
 		const [answer] = list;
@@ -40,7 +41,7 @@ client.on('interactionCreate', async interaction => {
 				{ name: 'Example', value: trim(answer.example, 1024) },
 				{ name: 'Rating', value: `${answer.thumbs_up} thumbs up. ${answer.thumbs_down} thumbs down.` },
 			);
-		interaction.reply({ embeds: [embed] });
+		interaction.editReply({ embeds: [embed] });
 	}
 });