+++ code-samples/sequelize/tags/sequelize.js
@@ -45,9 +45,9 @@ client.once('ready', () => {
 client.on('interactionCreate', async interaction => {
 	if (!interaction.isCommand()) return;
 
-	const { commandName: command } = interaction;
+	const { commandName } = interaction;
 
-	if (command === 'addtag') {
+	if (commandName === 'addtag') {
 		const tagName = interaction.options.getString('name');
 		const tagDescription = interaction.options.getString('description');
 
@@ -67,7 +67,7 @@ client.on('interactionCreate', async interaction => {
 
 			return interaction.reply('Something went wrong with adding a tag.');
 		}
-	} else if (command === 'tag') {
+	} else if (commandName === 'tag') {
 		const tagName = interaction.options.getString('name');
 
 		// equivalent to: SELECT * FROM tags WHERE name = 'tagName' LIMIT 1;
@@ -80,7 +80,7 @@ client.on('interactionCreate', async interaction => {
 		}
 
 		return interaction.reply(`Could not find tag: ${tagName}`);
-	} else if (command === 'edittag') {
+	} else if (commandName === 'edittag') {
 		const tagName = interaction.options.getString('name');
 		const tagDescription = interaction.options.getString('description');
 
@@ -92,7 +92,7 @@ client.on('interactionCreate', async interaction => {
 		}
 
 		return interaction.reply(`Could not find a tag with name ${tagName}.`);
-	} else if (command === 'taginfo') {
+	} else if (commandName === 'taginfo') {
 		const tagName = interaction.options.getString('name');
 
 		// equivalent to: SELECT * FROM tags WHERE name = 'tagName' LIMIT 1;
@@ -103,12 +103,12 @@ client.on('interactionCreate', async interaction => {
 		}
 
 		return interaction.reply(`Could not find tag: ${tagName}`);
-	} else if (command === 'showtags') {
+	} else if (commandName === 'showtags') {
 		const tagList = await Tags.findAll({ attributes: ['name'] });
 		const tagString = tagList.map(t => t.name).join(', ') || 'No tags set.';
 
 		return interaction.reply(`List of tags: ${tagString}`);
-	} else if (command === 'removetag') {
+	} else if (commandName === 'removetag') {
 		// equivalent to: DELETE from tags WHERE name = ?;
 		const tagName = interaction.options.getString('name');
 		const rowCount = await Tags.destroy({ where: { name: tagName } });