+++ guide/.vuepress/components/ResultingCode.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script setup lang="ts">
-import { defineProps, onMounted, ref } from 'vue';
+import { onMounted, ref } from 'vue';
 import { usePageData } from '@vuepress/client';
 
 const codeSamplesURL = 'https://github.com/discordjs/guide/tree/main/code-samples/';